strnstr fini et size_t remis dans chaque fichier

+++ ft_bzero.c
@@ -1,7 +1,7 @@
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*tmp;
-	unsigned int	i;
+	size_t	i;
 
 	tmp = s;
 	i = 0;
