+++ ft_memset.c
@@ -1,6 +1,6 @@
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	unsigned char	*tab;
 
 	tab = (unsigned char *)s;
